LayoutTests/js/Promise-then-without-callbacks.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description('Test Promise.');

window.jsTestIsAsync = true;

new Promise(function(r) { r.fulfill('hello'); })
  .then()
  .then(function(result) {
    window.result = result;
    shouldBeEqualToString('result', 'hello');
    finishJSTest();
  });

</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
